manually sync 3d-v1.0.0 from 3d (#4817)
