+++ package.json
@@ -11,8 +11,8 @@
   "scripts": {
     "server": "http-server . -p 8002",
     "build": "npm run build:min && npm run build-declaration",
-    "build:dev": "node ./rollup/out/build-engine-cli.js --sourcemap --platform=any --physics=cannon --destination=./bin/cocos-3d.dev.js",
-    "build:min": "node ./rollup/out/build-engine-cli.js --compress --sourcemap --platform=any --physics=cannon --destination=./bin/cocos3d-js.min.js",
+    "build:dev": "node ./rollup/out/build-engine-cli.js --sourcemap --platform=universal --physics=cannon --destination=./bin/cocos-3d.dev.js",
+    "build:min": "node ./rollup/out/build-engine-cli.js --compress --sourcemap --platform=universal --physics=cannon --destination=./bin/cocos3d-js.min.js",
     "build-declaration": "node ./scripts/generate-declarations/generate-declarations-cli.js",
     "build-rollup": "tsc --build rollup/tsconfig.json",
     "postinstall": "npm run build-rollup",
@@ -70,11 +70,12 @@
     "multipipe": "0.3.0",
     "persistify": "^1.1.1",
     "require-dir": "0.3.0",
-    "rollup": "^0.65.2",
+    "rollup": "^1.15.6",
     "rollup-plugin-babel": "^4.1.0-1",
     "rollup-plugin-commonjs": "^9.1.6",
     "rollup-plugin-json": "^3.1.0",
     "rollup-plugin-node-resolve": "^3.4.0",
+    "rollup-plugin-uglify": "^6.0.2",
     "rollup-watch": "^4.3.1",
     "tfig": "^1.0.20",
     "tslint": "^5.12.0",
@@ -83,7 +84,6 @@
     "typedoc-plugin-internal-external": "^2.0.2",
     "typedoc-plugin-localization": "github:cocos-creator/typedoc-plugin-localization#master",
     "typescript": "^3.4.4",
-    "uglify-js": "^3.4.9",
     "vconsole": "^3.3.0",
     "vinyl": "1.1.0",
     "vinyl-buffer": "1.0.0",
